All citizen sub-classes that were added to the citizens_detail class were broken down into separate files and then added to a directory inside the resources directory.
Qsystem.py was also modified to reflect these changes.
