+++ api/app/resources/citizen/citizen_left.py
