+++ api/app/resources/citizen/citizens_list.py
